dl2024-ex08-transformers-dl2024-sigma_learning/requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pytest==7.1.2
pycodestyle==2.8
scipy==1.8
scikit-learn==1.1
seaborn==0.11.2
regex==2019.3.12
absl-py==0.7.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2
